[Demangle] Separate DeclNameExtractor & Clang USR to Decl
