+++ lib/AST/ASTDemangler.cpp
@@ -24,6 +24,7 @@
 #include "swift/AST/ASTContext.h"
 #include "swift/AST/ClangModuleLoader.h"
 #include "swift/AST/Decl.h"
+#include "swift/AST/DeclNameExtractor.h"
 #include "swift/AST/GenericSignature.h"
 #include "swift/AST/Module.h"
 #include "swift/AST/NameLookup.h"
@@ -39,167 +40,21 @@
 #include "swift/Demangling/Demangler.h"
 #include "swift/Demangling/ManglingMacros.h"
 #include "llvm/ADT/StringSwitch.h"
+#include "clang/Tooling/Refactor/USRFinder.h"
 
 using namespace swift;
 
-class DeclNameExtractor {
-  ASTContext &Ctx;
-  
-public:
-  DeclNameExtractor(ASTContext &ctx) : Ctx(ctx) {}
+static Decl *getClangDeclFromUSR(ASTContext &ctx, StringRef usr) {
+  auto *clangLoader = ctx.getClangModuleLoader();
+  auto &clangCtx = clangLoader->getClangASTContext();
+  auto *clangDecl = clang::tooling::rename::getNamedDeclWithUSR(clangCtx, usr);
 
-  /// Extract a DeclName from a demangling node
-  DeclName extractDeclName(Node *node) {
-    if (node == nullptr)
-      return DeclName();
-    
-    switch (node->getKind()) {
-    case Node::Kind::Class:
-    case Node::Kind::Structure:
-    case Node::Kind::Enum:
-    case Node::Kind::Protocol:
-    case Node::Kind::TypeAlias:
-    case Node::Kind::OtherNominalType:
-    case Node::Kind::AssociatedType:
-    case Node::Kind::AssociatedTypeRef:
-    case Node::Kind::GenericTypeParamDecl:
-    case Node::Kind::Variable:
-    case Node::Kind::Macro:
-      return extractIdentifierName(node);
+  return clangLoader->importDeclDirectly(clangDecl);
+}
 
-    case Node::Kind::Constructor:
-    case Node::Kind::Allocator:
-      return DeclName(DeclBaseName::createConstructor());
-
-    case Node::Kind::Destructor:
-    case Node::Kind::Deallocator:
-    case Node::Kind::IsolatedDeallocator:
-      return DeclName(DeclBaseName::createDestructor());
-
-    case Node::Kind::Module:
-      return extractTextName(node);
-
-    case Node::Kind::Function:
-      return extractFunctionLikeName(node);
-
-    case Node::Kind::Subscript:
-      return extractFunctionLikeName(node);
-
-    default:
-      // For any other node types, we can't extract a meaningful name
-      return DeclName();
-    }
-
-  }
-
-private:
-  DeclName extractIdentifierName(Node *node) {
-    auto Identifier = node->getChild(1);
-    if (Identifier == nullptr ||
-        Identifier->getKind() != Node::Kind::Identifier)
-      return DeclName();
-    
-    return extractTextName(Identifier);
-  }
-
-  DeclName extractTextName(Node *node) {
-    if (!node->hasText())
-      return DeclName();
-    
-    auto identifier = Ctx.getIdentifier(node->getText());
-    return DeclName(identifier);
-  }
-
-  DeclName extractFunctionLikeName(Node *node) {
-    assert(node->getKind() == Node::Kind::Function ||
-           node->getKind() == Node::Kind::Subscript);
-    
-    DeclBaseName BaseName;
-    if (node->getKind() == Node::Kind::Function)
-      BaseName = extractIdentifierName(node).getBaseName();
-    else
-      BaseName = DeclBaseName::createSubscript();
-    
-    if (BaseName.empty())
-      return DeclName();
-    
-    // Location of LabelList node if present, otherwise a Type node.
-    unsigned LabelListIdx;
-    if (node->getKind() == Node::Kind::Function)
-      LabelListIdx = 2;
-    else
-      LabelListIdx = 1;
-    
-    auto *LabelsOrType = node->getChild(LabelListIdx);
-    assert(LabelsOrType != nullptr && (LabelsOrType->getKind() == Node::Kind::LabelList ||
-                                     LabelsOrType->getKind() == Node::Kind::Type));
-
-    SmallVector<Identifier, 4> ArgLabels;
-    if (LabelsOrType->getKind() == Node::Kind::LabelList)
-      extractArgLabelsFromLabelList(LabelsOrType, ArgLabels);
-    else
-      extractArgLabelsFromType(LabelsOrType, ArgLabels);
-
-    if (ArgLabels.empty())
-      return DeclName(BaseName);
-
-    return DeclName(Ctx, BaseName, ArgLabels);
-  }
-  
-  void extractArgLabelsFromLabelList(Node *LabelList,
-                                     SmallVectorImpl<Identifier> &ArgLabels) {
-    assert(LabelList->getKind() == Node::Kind::LabelList);
-    
-    for (unsigned i = 0; i < LabelList->getNumChildren(); ++i) {
-      auto *Label = LabelList->getChild(i);
-
-      assert(Label && (Label->getKind() == Node::Kind::Identifier ||
-                       Label->getKind() == Node::Kind::FirstElementMarker));
-
-      if (Label->getKind() == Node::Kind::Identifier) {
-        ArgLabels.push_back(Ctx.getIdentifier(Label->getText()));
-      } else {
-        ArgLabels.push_back(Identifier());
-      }
-    }
-  }
-  
-  void extractArgLabelsFromType(Node *Type, SmallVectorImpl<Identifier> &ArgLabels) {
-    auto ArgTuple = Type->findByKind(Demangle::Node::Kind::ArgumentTuple,
-                                     /*maxDepth=*/5);
-    if (ArgTuple == nullptr)
-      return;
-    
-    auto Params = ArgTuple->getFirstChild();
-    auto ParamsType = Params->getFirstChild();
-    if (ParamsType == nullptr)
-      return;
-    
-    if (ParamsType->getKind() != Demangle::Node::Kind::Tuple) {
-      // A single, unnamed parameter
-      ArgLabels.push_back(Identifier());
-      return;
-    }
-    
-    // More than one parameter are present
-    while (Params && Params->getFirstChild() &&
-           Params->getFirstChild()->getKind() != Node::Kind::TupleElement) {
-      Params = Params->getFirstChild();
-    }
-    
-    if (Params) {
-      for (size_t i = 0; i < Params->getNumChildren(); ++i) {
-        ArgLabels.push_back(Identifier());
-      }
-    }
-  }
-};
-
-
-Decl *swift::Demangle::getDeclFromUSR(ASTContext &ctx,
-                                      StringRef usr,
-                                      const DeclContext *lookupDC,
-                                      GenericSignature genericSig) {
+static Decl *getSwiftDeclFromUSR(ASTContext &ctx, StringRef usr,
+                                 const DeclContext *lookupDC,
+                                 GenericSignature genericSig) {
   if (!usr.starts_with("s:"))
     return nullptr;
 
@@ -211,6 +66,16 @@ Decl *swift::Demangle::getDeclFromUSR(ASTContext &ctx,
 
   ASTBuilder builder(ctx, genericSig);
   return builder.findDecl(node, usr, lookupDC);
+}
+
+Decl *swift::Demangle::getDeclFromUSR(ASTContext &ctx,
+                                      StringRef usr,
+                                      const DeclContext *lookupDC,
+                                      GenericSignature genericSig) {
+  if (usr.starts_with("c:"))
+    return getClangDeclFromUSR(ctx, usr);
+  
+  return getSwiftDeclFromUSR(ctx, usr, lookupDC, genericSig);
 }
 
 Type swift::Demangle::getTypeForMangling(ASTContext &ctx,
